move unit tests to lib
