+++ champ/lib/p2p/lib.rs
@@ -1,13 +1,9 @@
-mod main;
-
-use tokio::net::{TcpListener, TcpStream, ToSocketAddrs};
-use tokio::net::tcp::{OwnedReadHalf, OwnedWriteHalf};
-use tokio_util::codec::{FramedRead, FramedWrite, LengthDelimitedCodec};
-use tokio_stream::StreamExt;
 use bytes::{Bytes, BytesMut};
 use libp2p::futures::SinkExt;
-
-
+use tokio::net::tcp::{OwnedReadHalf, OwnedWriteHalf};
+use tokio::net::{TcpListener, TcpStream, ToSocketAddrs};
+use tokio_stream::StreamExt;
+use tokio_util::codec::{FramedRead, FramedWrite, LengthDelimitedCodec};
 
 struct Connection {
     framed_write: FramedWrite<OwnedWriteHalf, LengthDelimitedCodec>,
@@ -15,6 +11,7 @@ struct Connection {
 }
 
 impl Connection {
+    #[allow(dead_code)]
     //waits for an incoming connection on address
     pub async fn listen<T: ToSocketAddrs>(address: T) -> Result<Connection, Box<dyn std::error::Error>> {
         let listener = TcpListener::bind(address).await?;
@@ -33,11 +30,10 @@ impl Connection {
     //creates a Connection from the TcpStream,
     fn connection_from_stream(stream: TcpStream) -> Connection {
         let (read_half, write_half) = stream.into_split();
-        let connection = Connection {
+        Connection {
             framed_read: FramedRead::new(read_half, LengthDelimitedCodec::new()),
             framed_write: FramedWrite::new(write_half, LengthDelimitedCodec::new()),
-        };
-        connection
+        }
     }
 
     #[allow(dead_code)]
@@ -48,6 +44,7 @@ impl Connection {
         Ok(())
     }
 
+    #[allow(dead_code)]
     //waits for connection to be able to read one frame
     pub async fn read(&mut self) -> BytesMut {
         let mut buffer: BytesMut = Default::default(); //fix this
@@ -57,4 +54,83 @@ impl Connection {
         }
         buffer
     }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[tokio::test]
+    //test passes when TcpStream::connect yields a TcpStream value and therefore succeeds
+    async fn listen_for_connection() {
+        let address = "127.0.0.1:7890";
+        tokio::spawn(async move {
+            Connection::listen(address).await.expect("failed listening for the connection");
+        });
+        TcpStream::connect(address).await.expect("failed connecting");
+    }
+
+    #[tokio::test]
+    #[should_panic]
+    //test fails when TcpStream::connect yields a TcpStream value and therefore did connect
+    async fn listen_for_wrong_connection() {
+        let address = "127.0.0.1:7890";
+        let other_address = "127.0.0.1:7891";
+        tokio::spawn(async move {
+            Connection::listen(address).await.expect("failed listening for the connection");
+        });
+        TcpStream::connect(other_address).await.expect("failed connecting");
+    }
+
+    #[tokio::test]
+    async fn read_frame() {
+        //this needs to be a different address to not interfere with other tests
+        //trying to figure out how to prevent this from happening
+        let address = "127.0.0.1:7892";
+        let buffer: Bytes = Bytes::from_static(b"testdata");
+        let expected = buffer.clone();
+
+        let read_handle = tokio::spawn(async move {
+            let mut connection = Connection::listen(address).await.unwrap();
+            let result = connection.read().await;
+            result
+        });
+
+        let write_handle = tokio::spawn(async move {
+            let stream = TcpStream::connect(address).await.unwrap();
+            let mut framed = FramedWrite::new(stream, LengthDelimitedCodec::new());
+            framed.send(buffer.clone()).await.unwrap();
+            framed.flush().await.unwrap(); //this flush is probably unnecessary
+        });
+
+        let result = read_handle.await.unwrap();
+        write_handle.await.unwrap();
+
+        assert_eq!(expected, result.to_owned())
+    }
+
+    //#[tokio::test]
+    //async fn write_frame() {
+    //    let address = "127.0.0.1:7890";
+    //    let mut buffer: Bytes = Bytes::from_static(b"testdata");
+    //    let mut expected = buffer.clone();
+    //
+    //    let read_handle = tokio::spawn(async move {
+    //        let mut connection = Connection::listen(address).await.unwrap();
+    //        let result = connection.write().await;
+    //        result
+    //    });
+    //
+    //    let write_handle = tokio::spawn(async move {
+    //        let mut stream = TcpStream::connect(address).await.unwrap();
+    //        let mut framed = FramedWrite::new(stream, LengthDelimitedCodec::new());
+    //        let x = framed.send(buffer).await.unwrap();
+    //        framed.flush().await.unwrap(); //this flush is probably unnecessary
+    //    });
+    //
+    //    let result = read_handle.await.unwrap();
+    //    write_handle.await.unwrap();
+    //
+    //    assert_eq!(buffer, result.to_owned())
+    //}
 }
